requirements.txt
numpy
scipy
h5py
docopt
pyyaml
playsound
openpyxl
pathlib
argparse